# scripts/openapi-snapshot.json
{
 "components": {
  "schemas": {
...
   "Body_upload_report_api_v1_reports_upload_post": {
    "properties": {
     "file": {
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "type": "string"
     },
     "programme": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Programme"
     },
     "sender": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender"
     }
    },
    "required": [
     "file"
    ],
    "title": "Body_upload_report_api_v1_reports_upload_post",
    "type": "object"
   },
...
   "ExceptionSummaryOut": {
    "properties": {
     "aggregate": {
      "additionalProperties": true,
      "title": "Aggregate",
      "type": "object"
     },
     "completed_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completed At"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "narrative": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Narrative"
     },
     "narrative_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Narrative Error"
     },
     "narrative_model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Narrative Model"
     },
     "narrative_status": {
      "title": "Narrative Status",
      "type": "string"
     },
     "narrative_warning": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Narrative Warning"
     },
     "report_id": {
      "title": "Report Id",
      "type": "string"
     }
    },
    "required": [
     "id",
     "report_id",
     "narrative_status",
     "aggregate",
     "narrative",
     "narrative_model",
     "narrative_error",
     "narrative_warning",
     "created_at",
     "completed_at"
    ],
    "title": "ExceptionSummaryOut",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobOut": {
    "properties": {
     "attempts": {
      "title": "Attempts",
      "type": "integer"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "finished_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "heartbeat_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Heartbeat At"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "max_attempts": {
      "title": "Max Attempts",
      "type": "integer"
     },
     "metrics": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Metrics"
     },
     "stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stage"
     },
     "started_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "id",
     "kind",
     "status",
     "stage",
     "attempts",
     "max_attempts",
     "error_code",
     "error_message",
     "created_at",
     "started_at",
     "finished_at",
     "heartbeat_at"
    ],
    "title": "JobOut",
    "type": "object"
   },
   "LoginRequest": {
    "additionalProperties": false,
    "properties": {
     "email": {
      "maxLength": 320,
      "minLength": 3,
      "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
      "title": "Email",
      "type": "string"
     },
     "password": {
      "maxLength": 256,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     }
    },
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest",
    "type": "object"
   },
...
   "MeOut": {
    "properties": {
     "can_write": {
      "title": "Can Write",
      "type": "boolean"
     },
     "csrf_token": {
      "title": "Csrf Token",
      "type": "string"
     },
     "permissions": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Permissions",
      "type": "array"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "tenant": {
      "$ref": "#/components/schemas/TenantOut"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "required": [
     "user",
     "tenant",
     "role",
     "can_write",
     "csrf_token"
    ],
    "title": "MeOut",
    "type": "object"
   },
...
   "Page_ReportOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ReportOut"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     },
     "offset": {
      "title": "Offset",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[ReportOut]",
    "type": "object"
   },
   "ReportOut": {
    "properties": {
     "coverage_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Coverage Pct"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "expires_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "file_kind": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "File Kind"
     },
     "file_name": {
      "title": "File Name",
      "type": "string"
     },
     "file_size_bytes": {
      "title": "File Size Bytes",
      "type": "integer"
     },
     "grade": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "ingest_notes": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingest Notes"
     },
     "issues_found": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Issues Found"
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "processing_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processing Error"
     },
     "programme": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Programme"
     },
     "rows_processed": {
      "title": "Rows Processed",
      "type": "integer"
     },
     "rows_total": {
      "title": "Rows Total",
      "type": "integer"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "sender": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender"
     },
     "sheet_count_total": {
      "title": "Sheet Count Total",
      "type": "integer"
     },
     "source_channel": {
      "default": "upload",
      "title": "Source Channel",
      "type": "string"
     },
     "source_sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Sha256"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "updated_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     }
    },
    "required": [
     "id",
     "created_at",
     "file_name",
     "file_size_bytes",
     "sheet_count_total",
     "rows_processed",
     "rows_total",
     "coverage_pct",
     "grade",
     "score",
     "status"
    ],
    "title": "ReportOut",
    "type": "object"
   },
...
   "SignupRequest": {
    "additionalProperties": false,
    "properties": {
     "display_name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Display Name",
      "type": "string"
     },
     "email": {
      "maxLength": 320,
      "minLength": 3,
      "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
      "title": "Email",
      "type": "string"
     },
     "organisation": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Organisation",
      "type": "string"
     },
     "password": {
      "maxLength": 256,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     }
    },
    "required": [
     "email",
     "password",
     "display_name",
     "organisation"
    ],
    "title": "SignupRequest",
    "type": "object"
   },
...
   "TenantOut": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "org_type": {
      "default": "capacity_provider",
      "title": "Org Type",
      "type": "string"
     },
     "require_2fa": {
      "default": false,
      "title": "Require 2Fa",
      "type": "boolean"
     },
     "retention_days": {
      "title": "Retention Days",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "name",
     "retention_days"
    ],
    "title": "TenantOut",
    "type": "object"
   },
   "UserOut": {
    "properties": {
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "email": {
      "title": "Email",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     }
    },
    "required": [
     "id",
     "email",
     "display_name"
    ],
    "title": "UserOut",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "TrueBind API",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/login": {
   "post": {
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/api/v1/auth/signup": {
   "post": {
    "operationId": "signup_api_v1_auth_signup_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SignupRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Signup",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/api/v1/reports": {
   "get": {
    "operationId": "list_reports_api_v1_reports_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "maximum": 10000000,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ReportOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Reports",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v1/reports/upload": {
   "post": {
    "operationId": "upload_report_api_v1_reports_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_report_api_v1_reports_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Upload Report",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/exceptions/summary": {
   "get": {
    "operationId": "get_latest_exception_summary_api_v1_reports__report_id__exceptions_summary_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExceptionSummaryOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Latest Exception Summary",
    "tags": [
     "exception_summary"
    ]
   },
   "post": {
    "operationId": "create_exception_summary_api_v1_reports__report_id__exceptions_summary_post",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExceptionSummaryOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Exception Summary",
    "tags": [
     "exception_summary"
    ]
   }
  },
...
 }
}